// src/shared/generated/StateGameWindowRecord.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DisplayTargetRecord } from "./DisplayTargetRecord";
import type { GameWindowPlacementRecord } from "./GameWindowPlacementRecord";
import type { GameWindowTabRecord } from "./GameWindowTabRecord";

export type StateGameWindowRecord = { id: string, name: string, targetDisplay: DisplayTargetRecord, placement: GameWindowPlacementRecord, tabs: Array<GameWindowTabRecord>, activeTabId?: string, createdAt: string, updatedAt: string, };